refactor: relocate module templatesets
