--- bcs-app/backend/apis/templatesets/views.py
@@ -1,31 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) available.
-# Copyright (C) 2017-2019 THL A29 Limited, a Tencent company. All rights reserved.
-# Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://opensource.org/licenses/MIT
-#
-# Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
-# an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
-# specific language governing permissions and limitations under the License.
-#
-from rest_framework.response import Response
-
-from backend.apis.views import BaseAPIViewSet
-from backend.apps.configuration import models
-from backend.apps.configuration.serializers_new import ListTemplateSLZ
-from backend.templatesets.legacy_apps.configuration.showversion import views as show_version_view
-
-
-class TemplatesViewSet(BaseAPIViewSet):
-    def list_templates(self, request, project_id):
-        templates = models.Template.objects.filter(project_id=project_id)
-        serializer = ListTemplateSLZ(templates, many=True, context={"kind": request.project.kind})
-        template_list = serializer.data
-        return Response(template_list)
-
-
-class ShowVersionViewSet(BaseAPIViewSet, show_version_view.ShowVersionViewSet):
-    pass
