--- bcs-app/backend/apis/templatesets/yaml_mode/__init__.py
@@ -1,1 +1,0 @@
-# -*- coding: utf-8 -*-
