+++ bcs-app/backend/templatesets/open_apis/form_mode/release.py
