+++ bcs-app/backend/templatesets/open_apis/form_mode/serializers.py
@@ -16,12 +16,12 @@ from collections import OrderedDict
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 
-from backend.apps.configuration import models
-from backend.apps.configuration.constants import K8sResourceName
-from backend.apps.configuration.utils import to_bcs_res_name
-from backend.apps.instance.models import InstanceConfig, VersionInstance
-from backend.apps.instance.utils import validate_ns_by_tempalte_id
+from backend.templatesets.legacy_apps.configuration import models
+from backend.templatesets.legacy_apps.configuration.constants import K8sResourceName
 from backend.templatesets.legacy_apps.configuration.showversion.serializers import GetShowVersionSLZ
+from backend.templatesets.legacy_apps.configuration.utils import to_bcs_res_name
+from backend.templatesets.legacy_apps.instance.models import InstanceConfig, VersionInstance
+from backend.templatesets.legacy_apps.instance.utils import validate_ns_by_tempalte_id
 
 
 # TODO refactor validate_instance_entity
