+++ bcs-app/backend/templatesets/open_apis/yaml_mode/views.py
